cleanup: refactor item_name
