+++ src/data/item_name.rs
@@ -3,7 +3,4 @@ pub use main::generate;
 mod main;
 
 pub(crate) mod helpers;
-pub(crate) mod subtypes;
-
-#[cfg(test)]
-mod test;
+mod subtype;
