--- src/data/item_name/test.rs
@@ -1,12 +1,0 @@
-mod ammo;
-mod bag;
-mod chest;
-mod gem;
-mod jewelry;
-mod light_source;
-mod misc_object;
-mod misc_usable;
-mod ranged_weapon;
-mod spike;
-mod wearable_gem;
-mod wand;
