--- src/data/item_name/test/wand.rs
@@ -1,20 +1,0 @@
-use crate::data::item_name::generate;
-use crate::generate_item::template::WandTemplate;
-use crate::generate_item;
-
-#[test]
-fn test_wand_unidentified() {
-    let mut item = generate_item::generate(Box::new(WandTemplate::WandOfCloneMonster), 0);
-
-    item.set_identified(false);
-    assert_eq!(generate(&item), "Unknown wand");
-}
-
-#[test]
-fn test_wand_identified() {
-    let mut item = generate_item::generate(Box::new(WandTemplate::WandOfCloneMonster), 0);
-
-    item.set_identified(true);
-    item.p1 = 0;
-    assert_eq!(generate(&item), "Wand of Clone Monster (0 charges)");
-}
